S'tack fixes for editor rendering errors and foreach/repeat keywords

+++ src/net/ncplanner/plannerator/planner/s_tack/token/keyword/RepeatKeyword.java
@@ -12,7 +12,7 @@ public class RepeatKeyword extends Keyword{
     @Override
     public void run(Script script){
         StackMethod func = script.pop().asMethod();
-        long repeats = script.pop().asInt().getValue();
+        int repeats = (int)(long)script.pop().asInt().getValue();
         script.repeatSubscript(func.getValue(), repeats);
     }
 }
